Revised literal values, integration of XMLLiterals
XML serialisation of the Fresnel result: Literals are now stored in a
literal element instead of directly in the value element.

Parsed literals: The value sempfres:parsed has been added as a value
of fresnel:value. It causes a literal to be parsed if the datatype is
supported. In the XML serialization, this causes XMLLiterals to be
parsed and the resulting tree to be added to the Fresnel result tree.
More precisely, the set of nodes in the XMLLiteral is added as
children to a xmlliteral element which becomes the child of the
corresponding value element.

Example lens making use of this feature:
    :someFormat a fresnel:Format ;
        fresnel:propertyFormatDomain somens:someproperty ;
        fresnel:value sempfres:parsed .

There is also the value sempfres:parsedForcefullyAsXML which causes
the literal to be parsed as if it was an XMLLiteral even when it is
not.